remove username => use email instead rename LoginController => SecurityController
