+++ src/main/java/io/github/cristaling/swegg/backend/repositories/UserRepository.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 @Repository
 public interface UserRepository extends JpaRepository<User, UUID> {
 
-    User getUserByUsernameAndPassword(String username, String password);
-    User getUserByUsername(String username);
+    User getUserByEmailAndPassword(String email, String password);
+    User getUserByEmail(String email);
 
 }
