+++ src/main/java/io/github/cristaling/swegg/backend/service/RegisterService.java
@@ -4,7 +4,6 @@ import io.github.cristaling.swegg.backend.core.user.User;
 import io.github.cristaling.swegg.backend.core.user.UserData;
 import io.github.cristaling.swegg.backend.repositories.UserDataRepository;
 import io.github.cristaling.swegg.backend.repositories.UserRepository;
-import io.github.cristaling.swegg.backend.utils.DateUtils;
 import io.github.cristaling.swegg.backend.web.requests.RegisterRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,22 +19,27 @@ public class RegisterService {
         this.userDataRepository = userDataRepository;
     }
 
-
+    /**
+     *
+     * @param registerRequest contains all the data for a new user
+     * @return null-> if the email already exists
+     *         user-> if it was successful
+     *  Save a userData and a user in db
+     */
     public User registerUserAccount(RegisterRequest registerRequest) {
-        if(userRepository.getUserByUsername(registerRequest.getEmail())!=null){
+        if (userRepository.getUserByEmail(registerRequest.getEmail()) != null) {
             return null;
         }
         UserData userData = new UserData();
-        userData.setBirthDate(DateUtils.parseLocalDateFromString(registerRequest.getBirthDate()));
-        userData.setEmail(registerRequest.getEmail());
+        userData.setBirthDate(registerRequest.getBirthDate());
         userData.setFirstName(registerRequest.getFirstName());
         userData.setLastName(registerRequest.getLastName());
         userData.setTown(registerRequest.getTown());
         User user = new User();
-        user.setUsername(registerRequest.getEmail());
+        user.setEmail(registerRequest.getEmail());
         user.setPassword(registerRequest.getPassword());
         user.setUserData(userData);
-        userData.setUser(user );
+        userData.setUser(user);
         userRepository.save(user);
         userRepository.flush();
         userDataRepository.save(userData);
