+++ src/main/java/io/github/cristaling/swegg/backend/web/requests/LoginRequest.java
@@ -2,23 +2,23 @@ package io.github.cristaling.swegg.backend.web.requests;
 
 public class LoginRequest {
 
-	private String username;
+	private String email;
 	private String password;
 
 	public LoginRequest() {
 	}
 
-	public LoginRequest(String username, String password) {
-		this.username = username;
+	public LoginRequest(String email, String password) {
+		this.email = email;
 		this.password = password;
 	}
 
-	public String getUsername() {
-		return username;
+	public String getEmail() {
+		return email;
 	}
 
-	public void setUsername(String username) {
-		this.username = username;
+	public void setEmail(String email) {
+		this.email = email;
 	}
 
 	public String getPassword() {
